CypherStringInterpolator scala3 implementation
