+++ core/src/main/scala-3/neotypes/implicits/mappers/CompatibilityMappers.scala
@@ -1,0 +1,4 @@
+package neotypes
+package implicits.mappers
+
+trait CompatibilityMappers
